relocate get_arraylike_plugin
