+++ bioio/tests/test_bio_image.py
@@ -2,7 +2,12 @@
 # -*- coding: utf-8 -*-
 
 import pathlib
-from importlib.metadata import EntryPoint
+import sys
+
+if sys.version_info < (3, 10):
+    from importlib_metadata import EntryPoint
+else:
+    from importlib.metadata import EntryPoint
 
 import bioio_base as biob
 import numpy as np
